refactor(extension): split push-down input and service modules under 500-line limit (#226)
Resolve feature-review findings TS-1/TS-2: two source files exceeded the
hard 500-line limit in general-code-change.md after the feature work.

- Extract the Claude push-down resolver group from mcp-tool-inputs.ts
  (557 -> 486) into mcp-tool-inputs-push-down.ts; re-export from the
  original path so no consumer import changes.
- Extract the push-down CLI arg-vector construction from
  repo-automation-service.ts (507 -> 484) into
  repo-automation-service-push-down.ts following the existing
  repo-automation-service-*.ts convention; the method is now a thin
  delegator preserving the backward-compatible arg vector.

Behavior-preserving: no MCP schema change, no test changes, no new
dependencies. TypeScript toolchain green; coverage 95.87% line /
88.07% branch (no regression).
