+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/baseline/phase0-instructions-read.remediation.2026-06-24T23-08.md
@@ -1,0 +1,23 @@
+# Phase 0 — Policy Read Evidence (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+
+Policy Order:
+1. CLAUDE.md
+2. .claude/rules/general-code-change.md
+3. .claude/rules/general-unit-test.md
+4. .claude/rules/typescript.md
+5. .claude/rules/typescript-suppressions.md
+6. .claude/rules/quality-tiers.md
+7. .claude/rules/architecture-boundaries.md
+
+Files Read (in order):
+- CLAUDE.md
+- .claude/rules/general-code-change.md
+- .claude/rules/general-unit-test.md
+- .claude/rules/typescript.md
+- .claude/rules/typescript-suppressions.md
+- .claude/rules/quality-tiers.md
+- .claude/rules/architecture-boundaries.md
+
+Output Summary: All seven policy files confirmed present and read. Key constraints applicable to this remediation: 500-line hard file limit (general-code-change.md); kebab-case filenames, ES modules, strong typing, no `any`/`@ts-ignore`/`@ts-nocheck`/file-level eslint-disable (typescript.md, typescript-suppressions.md); coverage line >= 85%, branch >= 75%, no regression on changed lines (quality-tiers.md, general-unit-test.md); architecture boundary rules (architecture-boundaries.md).
