+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/qa-gates/file-size.2026-06-24T23-08.md
@@ -1,0 +1,13 @@
+# Final QA File Size (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: wc -l extensions/drm-copilot/src/mcp-tool-inputs.ts extensions/drm-copilot/src/repo-automation-service.ts extensions/drm-copilot/src/mcp-tool-inputs-push-down.ts extensions/drm-copilot/src/repo-automation-service-push-down.ts
+EXIT_CODE: 0
+
+Output Summary:
+- extensions/drm-copilot/src/mcp-tool-inputs.ts: 486 lines (reduced from 557; <= 500)
+- extensions/drm-copilot/src/repo-automation-service.ts: 484 lines (reduced from 507; <= 500)
+- extensions/drm-copilot/src/mcp-tool-inputs-push-down.ts: 82 lines (new module; <= 500)
+- extensions/drm-copilot/src/repo-automation-service-push-down.ts: 33 lines (new module; <= 500)
+
+All four files are at or below the 500-line hard limit. Both original targets reduced below 500; both new modules well below 500.
