+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/qa-gates/scope-confirmation.2026-06-24T23-08.md
@@ -1,0 +1,29 @@
+# Scope Confirmation (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: git status --porcelain; git diff --name-only ea94a068e0a071940858a0694c47e204244c09af -- extensions/drm-copilot/src; grep prohibited-pattern scan
+EXIT_CODE: 0
+
+## Source-code changes (working tree, extensions/drm-copilot/src)
+
+Exactly four files, all in scope:
+- M extensions/drm-copilot/src/mcp-tool-inputs.ts (reduced to 486 lines; added re-exports; exported asToolArgumentObject)
+- M extensions/drm-copilot/src/repo-automation-service.ts (reduced to 484 lines; thin delegation)
+- ?? extensions/drm-copilot/src/mcp-tool-inputs-push-down.ts (new module, 82 lines)
+- ?? extensions/drm-copilot/src/repo-automation-service-push-down.ts (new module, 33 lines)
+
+Note on `git diff` vs merge base: `git diff --name-only ea94a068... -- extensions/drm-copilot/src` also lists `mcp-repo-automation-tool-definitions.ts`, `mcp-tool-definitions.ts`, and `repo-automation-command-registration-admin.ts`. These were changed by prior feature work already committed on this branch (the push-down feature itself); they are NOT modified by this remediation. The authoritative scope check for this remediation is the working-tree status (`git status --porcelain`), which shows only the four in-scope source files above.
+
+## Prohibited-condition checks
+
+- No edits to `.claude/rules/**` or `.github/instructions/**` (working tree contains only feature-folder docs/evidence and the four source files).
+- MCP schema unchanged: `push_down_claude_customizations` inputSchema still declares `packs`, `csharp_variant`, `memory_mode` as optional properties (none added to a `required` array) with `additionalProperties: false` retained (mcp-repo-automation-tool-definitions.ts lines 124-153). This file was not modified by this remediation.
+- No `any` / `@ts-ignore` / `@ts-nocheck` / file-level eslint-disable introduced (grep scan on the four files returned NO_PROHIBITED_PATTERNS).
+- No new runtime dependency (package.json unmodified).
+- No temp files in tests (no test files changed; all 415 tests unchanged and passing).
+
+## Other working-tree entries (not source code, expected)
+
+Feature-folder documents and evidence artifacts under docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/ (spec.md, user-story.md, audit/review/remediation docs, and evidence/ artifacts authored during this workflow).
+
+Output Summary: Only the four in-scope `.ts` source files changed. None of the prohibited conditions are present. PASS.
