+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/qa-gates/ts-format.remediation.2026-06-24T23-08.md
@@ -1,0 +1,7 @@
+# Final QA TypeScript Format (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: npm run format (prettier --write), run from extensions/drm-copilot/
+EXIT_CODE: 0
+
+Output Summary: Prettier reported every source/test/config file as "unchanged". No file was reformatted on the final pass. Format gate is clean.
