+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/qa-gates/ts-test.remediation.2026-06-24T23-08.md
@@ -1,0 +1,16 @@
+# Final QA TypeScript Test + Coverage (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: npm run test -- --coverage (node run-jest.cjs --coverage), run from extensions/drm-copilot/
+EXIT_CODE: 0
+
+Output Summary:
+- Test Suites: 36 passed, 36 total
+- Tests: 415 passed, 415 total
+- Overall coverage: 95.87% lines, 88.07% branch, 95.87% statements, 96.17% functions
+- src/mcp-tool-inputs.ts: 93.2% lines, 91.66% branch
+- src/mcp-tool-inputs-push-down.ts: 97.56% lines, 96.55% branch
+- src/repo-automation-service.ts: 100% lines, 79.06% branch
+- src/repo-automation-service-push-down.ts: 100% lines, 100% branch
+
+All suites pass. Overall coverage exceeds policy thresholds (line >= 85%, branch >= 75%). Both new modules are well covered. See coverage-delta.2026-06-24T23-08.md for the no-regression comparison.
