+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/remediation-baseline/file-size.2026-06-24T23-08.md
@@ -1,0 +1,11 @@
+# Baseline File Size (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: wc -l extensions/drm-copilot/src/mcp-tool-inputs.ts extensions/drm-copilot/src/repo-automation-service.ts
+EXIT_CODE: 0
+
+Output Summary:
+- extensions/drm-copilot/src/mcp-tool-inputs.ts: 557 lines (above 500-line hard limit)
+- extensions/drm-copilot/src/repo-automation-service.ts: 507 lines (above 500-line hard limit)
+
+Both files exceed the 500-line hard limit defined in `.claude/rules/general-code-change.md`.
