+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/remediation-baseline/ts-format.2026-06-24T23-08.md
@@ -1,0 +1,7 @@
+# Baseline TypeScript Format (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: npm run format (prettier --write), run from extensions/drm-copilot/
+EXIT_CODE: 0
+
+Output Summary: Prettier reported all 78 source/test/config files as "unchanged". No file was reformatted. Format baseline is clean.
