+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/remediation-baseline/ts-lint.2026-06-24T23-08.md
@@ -1,0 +1,7 @@
+# Baseline TypeScript Lint (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: npm run lint (eslint --no-error-on-unmatched-pattern src test), run from extensions/drm-copilot/
+EXIT_CODE: 0
+
+Output Summary: ESLint completed with 0 errors and 0 warnings. Lint baseline is clean.
