+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/remediation-baseline/ts-test.2026-06-24T23-08.md
@@ -1,0 +1,14 @@
+# Baseline TypeScript Test + Coverage (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: npm run test -- --coverage (node run-jest.cjs --coverage), run from extensions/drm-copilot/
+EXIT_CODE: 0
+
+Output Summary:
+- Test Suites: 36 passed, 36 total
+- Tests: 415 passed, 415 total
+- Overall coverage: 95.86% lines, 88.05% branch, 95.86% statements, 96.13% functions
+- src/mcp-tool-inputs.ts: 93.71% lines, 93.18% branch
+- src/repo-automation-service.ts: 100% lines, 81.25% branch
+
+All suites pass. Coverage exceeds policy thresholds (line >= 85%, branch >= 75%). These are the authoritative baseline coverage values for the no-regression delta check in Phase 3.
