+++ docs/features/active/2026-06-24-push-down-language-packs-csharp-variant-226/evidence/remediation-baseline/ts-typecheck.2026-06-24T23-08.md
@@ -1,0 +1,7 @@
+# Baseline TypeScript Type-Check (Remediation #226)
+
+Timestamp: 2026-06-24T23-08
+Command: npm run typecheck (tsc -p ./ --noEmit), run from extensions/drm-copilot/
+EXIT_CODE: 0
+
+Output Summary: TypeScript compiler completed with 0 type errors. Type-check baseline is clean.
